% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindhtml.R
\name{bindhtml}
\alias{bindhtml}
\title{bindhtml}
\usage{
bindhtml(.path)
}
\value{
This function does not return a value but it binds dashboard.html
and a validation html.
}
\description{
This script looks for the most recent HTML validation report in a folder
by searching for files whose names start with validation_cfg.
It then builds a combined dashboard.html page. If no validation file exists,
the output contains only a single “Dashboard” tab showing the main dashboard
HTML file. If a validation file is found, the script reads both the dashboard
and validation HTML files and creates a two-tab interface: one tab for the
dashboard and another for the validation report.
}
\examples{
\dontrun{
bindhtml(.path = "path/to/run")
}

}
