% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{call_functions}
\alias{call_functions}
\title{VS Code–friendly wrapper to select run folders}
\usage{
call_functions(preselect = NULL)
}
\arguments{
\item{preselect}{Optional numeric vector (indexes) or character vector (folder names) to select subfolders.
If NULL, all subfolders are selected.}
}
\value{
A list of selected subfolder entries.
}
\description{
Runs \code{check_files_and_list_subfolders()} and
\code{list_subfolders()} and selects
one or more subfolders automatically.
}
