% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combinedhtml}
\alias{combinedhtml}
\title{Append the content of one HTML file into another and save the result as a new HTML file.
dashboard and validation_cfg}
\usage{
combinedhtml(.path)
}
\arguments{
\item{.path}{the path of the folder}
}
\value{
combined.html
}
\description{
Append the content of one HTML file into another and save the result as a new HTML file.
dashboard and validation_cfg
}
\examples{
\dontrun{
combinedhtml <- combinedhtml(reports)
}

}
