% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGDXtoMIF.R
\name{convertGDXtoMIF}
\alias{convertGDXtoMIF}
\title{Convert GDX to MIF Format}
\usage{
convertGDXtoMIF(path_gdx, regions, mif_name, scenario_name)
}
\arguments{
\item{path_gdx}{Character. Path to the input GDX file.}

\item{regions}{Character vector. A list of regions to include in the analysis.}

\item{mif_name}{Character. Name of the output MIF file to be created. If NULL the formatted content is returned}

\item{scenario_name}{Character. Name of the scenario to associate with the generated MIF file.}
}
\value{
This function either creates a MIF file or implicitly returns the formatted content
}
\description{
This function processes a GDX file for specified regions and generates a MIF
report. The output combines several metrics into a single MIF file.
}
\details{
This function aggregates various metrics such as final energy, emissions,
socio-economic data, and more. The data is read from the provided GDX file
and written to a MIF file in a format compatible with the OPEN-PROM model.

The generated MIF file includes the following metrics:
\itemize{
  \item Final Energy
  \item Emissions
  \item Secondary Energy (SE)
  \item Primary Energy (PE)
  \item Gross Domestic Product (GDP)
  \item Population (POP)
  \item Carbon Prices
  \item Price Data
  \item Electricity Capacity
}
}
\examples{
\dontrun{
convertGDXtoMIF(
  path_gdx = "path/to/data.gdx",
  regions = c("MEA", "USA"),
  mif_name = "output.mif",
  scenario_name = "Scenario_name"
)
}
}
\seealso{
\code{\link[magclass]{write.report}}
}
