% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEmissions.r
\name{generateEmissions}
\alias{generateEmissions}
\title{Generate a structured CSV file for emissions data from a hardcoded list.}
\usage{
generateEmissions(
  allvariables,
  output_csv_path,
  sheet_name = "Emissions for MAGICC",
  start_year = 2015,
  end_year = 2100,
  region = "World",
  model = "unspecified",
  scenario = "unspecified"
)
}
\arguments{
\item{output_csv_path}{Character. The file path for the generated CSV file.}

\item{start_year}{Integer. The first year for the time series columns. Defaults to 1750.}

\item{end_year}{Integer. The last year for the time series columns. Defaults to 2500.}

\item{region}{Character. The region to specify in the output file. Defaults to "World".}

\item{model}{Character. The model name to specify. Defaults to "unspecified".}

\item{scenario}{Character. The scenario name to specify. Defaults to "unspecified".}
}
\value{
This function does not return a value but writes a CSV file to the specified path.
}
\description{
Generate a structured CSV file for emissions data from a hardcoded list.
}
\examples{
# generate_emissions_csv(output_csv_path = "emissions_data_template.csv")

}
