% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEmissionsFile.r
\name{generateEmissionsFile}
\alias{generateEmissionsFile}
\title{Generate Emissions CSV file for using it with climate-assessment tool of IIASA}
\usage{
generateEmissionsFile(
  .path,
  openPromVariables,
  years = NULL,
  scenario_name = NULL
)
}
\arguments{
\item{openPromVariables}{A magpie object that contains OPEN-PROM results.}

\item{years}{Optional; a character vector specifying the years. Defaults to the basename of `.path`.}

\item{scenario_name}{Optional; a character vector specifying the scenario names. Defaults to the basename of `.path`.}
}
\value{
This function does not return a value but writes a CSV file to the specified path.
}
\description{
This function processes the final output of the mif file and filters the required
variables for the emissions csv file. The emissions csv file is then used by the 
climate-assessment tool of IIASA. Only OPEN-PROM output data are used. 
It reads values from magpie object and avoid reading exported file.
}
\examples{
\dontrun{
generateEmissionsFile(.path = "path/to/gdx2", reports,
                years = c("y2015", "y2025"),
s               scenario_name = "NPI")
}

}
