% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_script_directory}
\alias{get_script_directory}
\title{Get the directory of the currently running script}
\usage{
get_script_directory()
}
\value{
Character path of the script directory
}
\description{
Works in Rscript, RStudio, and interactive sessions.
}
