% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlReportValidation.R
\name{htmlReportValidation}
\alias{htmlReportValidation}
\title{Creates a validation report file inn html with piamValidation}
\usage{
htmlReportValidation(.path, openPromFile)
}
\arguments{
\item{openPromFile}{A mif file that contains OPEN-PROM results.}
}
\value{
This function does not return a value but it creates a validation html file with piamValidation.
}
\description{
Creates a validation report file inn html with piamValidation
}
\examples{
\dontrun{
htmlReportValidation(.path = "path/to/run", "path/to/mif")
}

}
