% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReport.R
\name{plotReport}
\alias{plotReport}
\title{Generate Plots for MAgPIE Data}
\usage{
plotReport(
  magpie_obj,
  an,
  vars,
  plot_type = "bar",
  label = NULL,
  save_name = NULL
)
}
\arguments{
\item{magpie_obj}{A MAgPIE object containing the data to be plotted.}

\item{an}{A numeric vector specifying the analysis years.}

\item{vars}{A character vector or list of character vectors indicating the variables to plot.}

\item{plot_type}{A string specifying the plot type. Options are `"bar"` or `"area"`. Default is `"bar"`.}

\item{label}{An optional string to label the plot legend. Defaults to `NULL`.}

\item{save_name}{An optional string specifying the file name to save the plot. Defaults to `NULL`.}
}
\value{
A single plot object or a list of plot objects.
}
\description{
This function generates plots for MAgPIE objects based on specified variables and plot types.
}
\examples{
\dontrun{
plotReport(magpie_obj, an = c(2020, 2030), vars = "Capacity|Electricity", plot_type = "bar")
}
}
