% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read Data File}
\usage{
readData(.path, ...)
}
\arguments{
\item{.path}{The file path to the data file (string).}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class "magpie", or NULL if an error occurs.
}
\description{
A generic function to read data files.
Methods are provided for \code{.gdx} and \code{.mif} file types.
}
\examples{
data <- readData(system.file("extdata", "blabla.gdx", package = "openprom"), "iSet")
}
\seealso{
\link[gdx:readGDX]{gdx::readGDX} which this function wraps.
}
