% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlotReport.R
\name{reportAreaPNG}
\alias{reportAreaPNG}
\title{Create PNG area plots for the report}
\usage{
reportAreaPNG(report, grouped, output_dir)
}
\arguments{
\item{report}{MAgPIE report object.}

\item{grouped}{Data frame containing variable grouping information.}

\item{output_dir}{Character string specifying the parent output directory.}
}
\value{
Invisibly returns NULL.
}
\description{
Create PNG area plots for the report
}
\author{
Fotis Sioutas
}
