% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBudget.R
\name{reportBudget}
\alias{reportBudget}
\title{Plot Carbon Budget}
\usage{
reportBudget(
  magpieObject,
  aggregate = TRUE,
  budgetBaseYear = 2019,
  budget1p5 = 400,
  budget2c = 1150,
  probLabel = "67\%"
)
}
\arguments{
\item{budgetBaseYear}{The cumulated CO2 emissions in the year before 2020 where the budget begins.}

\item{budget1p5}{The carbon budget for CO2 emissions in Gt CO2 reaching 1.5C until 2100.}

\item{budget2c}{The carbon budget for CO2 emissions in Gt CO2 reaching 2C until 2100.}

\item{probLabel}{The probability for the carbon budgets.}
}
\value{
A magpie object containing carbon budgets.
}
\description{
This function adds two variables in the magpie object 
in order to plot two vertical lines with the budgets.
}
\examples{
\dontrun{
result <- reportBudget(magpieObject=example_magpie,aggregate= TRUE, budgetBaseYear = 2019,
                       budget1p5= 400,budget2c= 1150,probLabel= "67\%")}

}
