% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEfficiency.R
\name{reportEfficiency}
\alias{reportEfficiency}
\title{Energy Efficiency (GDP/TFC) 
Energy intensity (TFC/GDP)}
\usage{
reportEfficiency(reports, path, regions, years)
}
\arguments{
\item{reports}{Magpie object created from postprom.}
}
\value{
A magpie object containing Efficiency.
}
\description{
Energy Efficiency (GDP/TFC) 
Energy intensity (TFC/GDP)
}
\examples{
\dontrun{
result <- reportEfficiency(reports)
}

}
