% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmissionsMagpie.R
\name{reportEmissionsMagpie}
\alias{reportEmissionsMagpie}
\title{Process and Aggregate CO2 Emissions Data}
\usage{
reportEmissionsMagpie(path, regions, years)
}
\arguments{
\item{path}{The file path to the mif data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed and aggregated emissions data.
}
\description{
This function processes and aggregates emissions data from a mif file of
Magpie.
}
\examples{
\dontrun{
result <- reportEmissionsMagpie(system.file("extdata", "blabla.gdx", package = "postprom"), c("MEA"))
}

}
