% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFuelConsTargets.R
\name{reportFuelConsTargets}
\alias{reportFuelConsTargets}
\title{Process and Aggregate Fuel Consumption targets}
\usage{
reportFuelConsTargets(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed and aggregated Fuel Consumption.
}
\description{
This function processes and aggregates Fuel Consumption targets from csv files in folder data.
Is including final consumption for TRANSE,NENSE,INDSE,DOMSE sectors.
}
\examples{
\dontrun{
result <- reportFuelConsTargets(system.file("extdata", "blabla.gdx", package = "openprom"), c("MEA"))
}

}
