% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrossInlandConsumption.R
\name{reportGrossInlandConsumption}
\alias{reportGrossInlandConsumption}
\title{Process and Aggregate Gross Inland Consumption|Biomass and Waste}
\usage{
reportGrossInlandConsumption(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed carbon price data with proper units.
}
\description{
This function processes and aggregates Gross Inland Consumption|Biomass and Waste data from a GDX file,
converting it to a standardized format and attaching proper units.
}
\examples{
\dontrun{
result <- reportGrossInlandConsumption(system.file("extdata", "blabla.gdx", package = "postprom"), c("MEA"))
}

}
