% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrowthRates.R
\name{reportGrowthRates}
\alias{reportGrowthRates}
\title{Growth Rates of Emissions}
\usage{
reportGrowthRates(reports)
}
\arguments{
\item{reports}{Magpie object created from postprom.}
}
\value{
A magpie object containing Growth Rates.
}
\description{
This function computes Growth Rates of Emissions, Gross Emissions,
Carbon Capture and Carbon Removal.
}
\examples{
\dontrun{
result <- reportGrowthRates(reports)
}

}
