% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIndicators.R
\name{reportIndicators}
\alias{reportIndicators}
\title{Energy Efficiency (GDP/TFC) 
Energy intensity (TFC/GDP)}
\usage{
reportIndicators(reports, path, regions, years, blabla_regions)
}
\arguments{
\item{reports}{Magpie object created from postprom.}
}
\value{
A magpie object containing Indicators
}
\description{
Energy Efficiency (GDP/TFC) 
Energy intensity (TFC/GDP)
}
\examples{
\dontrun{
result <- reportIndicators(reports)
}

}
