% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPE.R
\name{reportPE}
\alias{reportPE}
\title{Process and Aggregate Primary Energy Data}
\usage{
reportPE(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed and aggregated primary energy data.
}
\description{
A function to process and aggregate primary energy data from a GDX file.
}
\examples{
\dontrun{
result <- reportPE(system.file("extdata", "blabla.gdx", package = "openprom"), c("MEA"))
}
}
