% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPOP.R
\name{reportPOP}
\alias{reportPOP}
\title{Process and Aggregate Population Data}
\usage{
reportPOP(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed population data with proper units.
}
\description{
This function processes and aggregates population data from a GDX file,
converting it to a standardized format and attaching proper units.
}
\examples{
\dontrun{
result <- reportPOP(system.file("extdata", "blabla.gdx", package = "postprom"), c("MEA"))
}

}
