% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceCarbon.R
\name{reportPriceCarbon}
\alias{reportPriceCarbon}
\title{Process and Aggregate Carbon Price Data}
\usage{
reportPriceCarbon(path, regions)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed carbon price data with proper units.
}
\description{
This function processes and aggregates carbon price data from a GDX file,
converting it to a standardized format and attaching proper units.
}
\examples{
\dontrun{
result <- reportPriceCarbon(system.file("extdata", "blabla.gdx", package = "openprom"), c("MEA"))
}

}
