% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSE.R
\name{reportSE}
\alias{reportSE}
\title{Process and Aggregate Secondary Energy Data}
\usage{
reportSE(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed and aggregated secondary energy data.
}
\description{
This function processes and aggregates secondary energy data.
It maps fuel categories, aggregates data by reporting categories, and formats the results into a magpie object.
}
\examples{
\dontrun{
result <- reportSE(system.file("extdata", "blabla.gdx", package = "openprom"), c("MEA"))
}

}
