% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationMif.R
\name{ValidationMif}
\alias{ValidationMif}
\title{Convert mif file to magpie object.}
\usage{
ValidationMif(
  .path,
  mif_name = "fullValidation2.mif",
  Validation_data_for_plots = Validation_data_for_plots,
  reportOPEN_PROM = reports,
  metadata_run = metadata_run,
  Validation2050 = Validation2050
)
}
\arguments{
\item{.path}{A character vector specifying the path(s) to the GDX files.}

\item{mif_name}{A string specifying the name of the MIF file to create.}

\item{Validation_data_for_plots}{Optional; a logical value indicating whether to add to plots validation data. Defaults to `TRUE`.}
}
\value{
A magpie object provided by validation mif.
}
\description{
This function processes validation mif file and converts to magpie object.
}
\examples{
\dontrun{
ValidationMif(.path = c("path/to/gdx1"),
mif_name = "output.mif",Validation_data_for_plots = TRUE)
}
}
