% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReport.R
\name{plotReport}
\alias{plotReport}
\title{Generate a Plot from a MAgPIE Object}
\usage{
plotReport(magpie_obj, plot_type = "bar", label = NULL, save_name = NULL)
}
\arguments{
\item{magpie_obj}{A MAgPIE object containing the data to be plotted.}

\item{plot_type}{A string specifying the type of plot. Options are `"bar"` or `"area"`. Defaults to `"bar"`.}

\item{label}{An optional string to label the plot legend. Defaults to `NULL`.}

\item{save_name}{An optional string specifying the file name to save the plot. If `NULL`, the plot is not saved. Defaults to `NULL`.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function generates a plot for a MAgPIE object using specified settings. It supports both bar and area plot types and handles color mappings for variables.
}
\details{
- The function extracts variable names from the `magpie_obj` and matches them with color mappings.
- A warning is issued if any variables do not have corresponding color mappings.
- If `save_name` is provided, the plot is saved to the specified file in high resolution.
}
\examples{
\dontrun{
plotReport(magpie_obj = example_magpie, plot_type = "bar", save_name = "example_plot.png")
}
}
