% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReport.R
\name{plotReport}
\alias{plotReport}
\title{Generate and Save Plots for MAGPIE Objects}
\usage{
plotReport(obj, ...)
}
\description{
This set of functions generates plots for MAGPIE objects or lists
of MAGPIE objects, based on the specified parameters.
The plots can be saved individually or combined into a PDF file.
}
\details{
The generic \code{plotReport} function delegates to specific methods based on the class of \code{obj}. Supported
methods include \code{plotReport.list} for lists of MAGPIE objects and \code{plotReport.magpie} for individual
MAGPIE objects.
}
\examples{
\dontrun{
magpie_obj <- read.magpie("example_file.mgp")
plotReport(magpie_obj, an = 2050, vars = c("variable1", "variable2"), plot_type = "bar")
}
}
