% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlotReport.R
\name{reportAreaPNG}
\alias{reportAreaPNG}
\title{Create PNG area plots for the report}
\usage{
reportAreaPNG(report, grouped, plotstyle, regionsPNG, yearsPNG, output_dir)
}
\arguments{
\item{report}{MAgPIE report object.}

\item{grouped}{Data frame containing variable grouping information.}

\item{plotstyle}{Data frame containing colors for each variable.}

\item{regionsPNG}{Character vector of regions to plot.}

\item{yearsPNG}{Numeric vector of years to plot.}

\item{output_dir}{Output directory for the PNG files.}
}
\value{
Invisibly returns NULL.
}
\description{
Create PNG area plots for the report
}
\author{
Fotis Sioutas
}
