% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBudget.R
\name{reportBudget}
\alias{reportBudget}
\title{Plot Carbon Budget}
\usage{
reportBudget(
  magpieObject,
  aggregate = TRUE,
  budgetBaseYear = 2019,
  budget1p5 = 400,
  budget2c = 1150,
  probLabel = "67\%"
)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed population data with proper units.
}
\description{
This function adds two variables in the magpie object 
in order to plot two vertical lines with the budgets.
}
\examples{
\dontrun{
result <- reportBudget(magpieObject=example_magpie,aggregate= TRUE, budgetBaseYear = 2019,
                       budget1p5= 400,budget2c= 1150,probLabel= "67\%")}

}
