% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEquipmentCapacityShare.R
\name{reportEquipmentCapacityShare}
\alias{reportEquipmentCapacityShare}
\title{Equipment Capacity Share.}
\usage{
reportEquipmentCapacityShare(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing Equipment Capacity shares data with proper units.
}
\description{
This function plots Equipment Capacity shares.
}
\examples{
\dontrun{
result <- reportEquipmentCapacityShare(system.file("extdata", "blabla.gdx", package = "postprom"), c("MEA"))
}

}
