% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTrade.R
\name{reportTrade}
\alias{reportTrade}
\title{Process and Aggregate Energy Trade Data}
\usage{
reportTrade(path, regions, years)
}
\arguments{
\item{path}{The file path to the GDX data file.}

\item{regions}{A character vector of region names to filter data.}
}
\value{
A magpie object containing processed and aggregated trade data with proper units.
}
\description{
This function processes gross imports, gross exports and net exports of all fuels
from a GDX file and aggregates them to every fuel category of the BALEFtoEF mapping.
}
\examples{
\dontrun{
result <- reportTrade(system.file("extdata", "blabla.gdx", package = "postprom"), c("MEA"))
}

}
